clean code remove not to used in any code
