+++ pom.xml
@@ -66,6 +66,11 @@
             <version>RELEASE</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>commons-codec</groupId>
+            <artifactId>commons-codec</artifactId>
+            <version>1.16.1</version> <!-- atau versi terbaru -->
+        </dependency>
     </dependencies>
 
 	<build>
