+++ src/main/java/com/core/payment_gateway_service/controller/PaymentGatewayAdminController.java
@@ -1,12 +1,18 @@
-package com.core.payment_gateway_service.Controller;
+package com.core.payment_gateway_service.controller;
 
-import com.core.payment_gateway_service.DTO.PaymentGatewayConfigRequest;
-import com.core.payment_gateway_service.DTO.PaymentGatewayConfigResponse;
-import com.core.payment_gateway_service.DTO.PaymentGatewayRequest;
-import com.core.payment_gateway_service.DTO.PaymentGatewayResponse;
+import com.core.payment_gateway_service.dto.PaymentGatewayConfigRequest;
+import com.core.payment_gateway_service.dto.PaymentGatewayConfigResponse;
+import com.core.payment_gateway_service.dto.PaymentGatewayRequest;
+import com.core.payment_gateway_service.dto.PaymentGatewayResponse;
 import com.core.payment_gateway_service.service.PaymentGatewayService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
