+++ src/main/java/com/core/payment_gateway_service/dto/BillResponse.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
