+++ src/main/java/com/core/payment_gateway_service/dto/FlipResponse.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 
 import com.core.payment_gateway_service.entity.PaymentGatewayTransaction;
 import com.fasterxml.jackson.annotation.JsonFormat;
