+++ src/main/java/com/core/payment_gateway_service/dto/ObjectUtil.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 import lombok.AccessLevel;
 
 import lombok.NoArgsConstructor;
