+++ src/main/java/com/core/payment_gateway_service/dto/PaymentGatewayResponse.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,11 +11,10 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class PaymentGatewayConfigResponse {
-    private String pgConfigId;
-    private String configName;
-    private String configValue; // bisa disensor nanti kalau `isSensitive` = true
-    private Boolean isSensitive;
+public class PaymentGatewayResponse {
+    private String paymentGatewayId;
+    private String gatewayName;
+    private String gatewayCode;
     private String description;
     private Boolean isActive;
     private Timestamp createdAt;
