+++ src/main/java/com/core/payment_gateway_service/dto/PaymentGatewayTransactionRequest.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 
 import com.core.payment_gateway_service.entity.EscrowAccount;
 import com.core.payment_gateway_service.entity.LoanAccount;
