+++ src/main/java/com/core/payment_gateway_service/dto/SpecificationHelper.java
@@ -1,4 +1,4 @@
-package com.core.payment_gateway_service.DTO;
+package com.core.payment_gateway_service.dto;
 
 import com.core.payment_gateway_service.enums.SearchOperator;
 import jakarta.persistence.criteria.Path;
