+++ src/main/java/com/core/payment_gateway_service/entity/EscrowAccount.java
@@ -1,10 +1,17 @@
 package com.core.payment_gateway_service.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-import java.math.BigDecimal;
-import java.sql.Timestamp;
 
 @Entity
 @Table(name = "escrow_accounts")
