+++ src/main/java/com/core/payment_gateway_service/entity/PaymentGatewayCallback.java
@@ -1,18 +1,23 @@
 package com.core.payment_gateway_service.entity;
 
 import com.core.payment_gateway_service.enums.CallbackProcessingStatus;
-import com.core.payment_gateway_service.service.JsonMapConverter;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.GenericGenerator;
-
 import java.sql.Timestamp;
-import java.util.Map;
-import java.util.UUID;
+
 
 @Entity
 @Table(name = "payment_gateway_callbacks")
