+++ src/main/java/com/core/payment_gateway_service/repository/PaymentGatewayTransactionRepository.java
@@ -3,7 +3,5 @@ package com.core.payment_gateway_service.repository;
 import com.core.payment_gateway_service.entity.PaymentGatewayTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.UUID;
-
 public interface PaymentGatewayTransactionRepository extends JpaRepository<PaymentGatewayTransaction, String> {
 }
