+++ src/main/java/com/core/payment_gateway_service/service/PaymentFlip.java
@@ -1,12 +1,25 @@
 package com.core.payment_gateway_service.service;
 
-import com.core.payment_gateway_service.DTO.*;
 import com.core.payment_gateway_service.config.FlipConfiguration;
-import com.core.payment_gateway_service.entity.*;
+import com.core.payment_gateway_service.dto.BilPaymentSearch;
+import com.core.payment_gateway_service.dto.BillPaymentRequest;
+import com.core.payment_gateway_service.dto.BillPaymentResponse;
+import com.core.payment_gateway_service.dto.FlipResponse;
+import com.core.payment_gateway_service.dto.SpecificationHelper;
+import com.core.payment_gateway_service.entity.BillPayment;
+import com.core.payment_gateway_service.entity.PaymentGatewayCallback;
+import com.core.payment_gateway_service.entity.PaymentGatewayTransaction;
 import com.core.payment_gateway_service.enums.CallbackProcessingStatus;
 import com.core.payment_gateway_service.enums.TransactionStatus;
 import com.core.payment_gateway_service.enums.TransactionType;
-import com.core.payment_gateway_service.repository.*;
+import com.core.payment_gateway_service.repository.BillPaymentRepository;
+import com.core.payment_gateway_service.repository.EscrowAccountRepository;
+import com.core.payment_gateway_service.repository.LoanAccountRepository;
+import com.core.payment_gateway_service.repository.PaymentGatewayCallbackRepository;
+import com.core.payment_gateway_service.repository.PaymentGatewayConfigRepository;
+import com.core.payment_gateway_service.repository.PaymentGatewayRepository;
+import com.core.payment_gateway_service.repository.PaymentGatewayTransactionRepository;
+import com.core.payment_gateway_service.repository.SavingAccountRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -14,7 +27,6 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -27,7 +39,6 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
-import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.time.LocalDateTime;
