+++ src/main/java/com/core/payment_gateway_service/service/PaymentGatewayServiceImpl.java
@@ -1,9 +1,9 @@
 package com.core.payment_gateway_service.service;
 
-import com.core.payment_gateway_service.DTO.PaymentGatewayConfigRequest;
-import com.core.payment_gateway_service.DTO.PaymentGatewayConfigResponse;
-import com.core.payment_gateway_service.DTO.PaymentGatewayRequest;
-import com.core.payment_gateway_service.DTO.PaymentGatewayResponse;
+import com.core.payment_gateway_service.dto.PaymentGatewayConfigRequest;
+import com.core.payment_gateway_service.dto.PaymentGatewayConfigResponse;
+import com.core.payment_gateway_service.dto.PaymentGatewayRequest;
+import com.core.payment_gateway_service.dto.PaymentGatewayResponse;
 import com.core.payment_gateway_service.entity.PaymentGateway;
 import com.core.payment_gateway_service.entity.PaymentGatewayConfig;
 import com.core.payment_gateway_service.repository.PaymentGatewayConfigRepository;
